Record which person wrote a message
`role` said a human spoke; nothing said which one. That was answerable by
inference while a conversation held one person -- every user message was the
owner's -- and stops being answerable the moment a second person is in one.

Nullable with no backfill, deliberately. Every existing user message belongs
to the conversation's owner, so a backfill would be derivable rather than
informative, and it would write the owner's id onto surface messages from
somebody with no account here at all.

The message value objects move to `domain.messages` and are re-exported,
because `value_objects` reached the per-file limit. Every caller still
imports them from where it did.
